+++ Software/MATLAB/app/system/+aws/+dynamodbv2/+document/@BatchWriteItemOutcome/getBatchWriteItemResult.m
@@ -1,0 +1,11 @@
+function result = getBatchWriteItemResult(obj)
+% GETBATCHWRITEITEMRESULT Returns a non-null low-level result
+% A BatchWriteItemResult object is returned.
+    
+% Copyright 2020 The MathWorks, Inc.
+       
+resultJ = obj.Handle.getBatchWriteItemResult();
+result = aws.dynamodbv2.model.BatchWriteItemResult(resultJ);
+
+end
+