+++ Software/MATLAB/app/system/+aws/+dynamodbv2/+model/@AttributeValue/setNS.m
@@ -1,8 +1,11 @@
 function setNS(obj, S)
 % SETNS Set a cell array of attributes of type Number
-% Cell array entries should be of type character vector
+% Cell array entries should be of type character vector.
+% The cell array will be mapped to a 1D array as it is converted to a Java List.
+% Dimensional structure will not be retained.
+% Insert order will not be retained.
 
-% Copyright 2019 The MathWorks, Inc.
+% Copyright 2020 The MathWorks, Inc.
 
 if ~iscellstr(S) %#ok<ISCLSTR>
     % Create logger reference