+++ Software/MATLAB/app/system/+aws/+dynamodbv2/@AmazonDynamoDBClient/AmazonDynamoDBClient.m
@@ -1,5 +1,5 @@
 classdef AmazonDynamoDBClient < aws.Object
-    % CLIENT Object to represent an AWS DynamoDB client
+    % CLIENT Object to represent an Amazon DynamoDB client
     % The client is used to carry out operations with the DynamoDB service
     %
     % Example:
@@ -14,7 +14,7 @@ classdef AmazonDynamoDBClient < aws.Object
     %    % Shutdown the client when no longer needed
     %    ddb.shutdown();
 
-    % Copyright 2018 The MathWorks, Inc.
+    % Copyright 2018-2021 The MathWorks, Inc.
 
     properties
         % default to using the AWS provider chain
@@ -36,7 +36,7 @@ classdef AmazonDynamoDBClient < aws.Object
             write(logObj,'verbose','Creating Client');
             % error if JVM is not enabled or MATLAB is too old
             if ~usejava('jvm')
-                write(logObj,'error','MATLAB must be used with the JVM enabled to access AWS DynamoDB');
+                write(logObj,'error','MATLAB must be used with the JVM enabled to access Amazon DynamoDB');
             end
             if verLessThan('matlab','9.2') % R2017a
                 write(logObj,'error','MATLAB Release 2017a or newer is required');